[maven-release-plugin]  copy for tag v8

+++ service/mcs/persistent/src/main/java/eu/europeana/cloud/service/mcs/persistent/CassandraDataSetService.java
@@ -8,10 +8,7 @@ import eu.europeana.cloud.common.response.CloudTagsResponse;
 import eu.europeana.cloud.common.response.ResultSlice;
 import eu.europeana.cloud.service.mcs.DataSetService;
 import eu.europeana.cloud.service.mcs.UISClientHandler;
-import eu.europeana.cloud.service.mcs.exception.DataSetAlreadyExistsException;
-import eu.europeana.cloud.service.mcs.exception.DataSetNotExistsException;
-import eu.europeana.cloud.service.mcs.exception.ProviderNotExistsException;
-import eu.europeana.cloud.service.mcs.exception.RepresentationNotExistsException;
+import eu.europeana.cloud.service.mcs.exception.*;
 import eu.europeana.cloud.service.mcs.persistent.cassandra.CassandraDataSetDAO;
 import eu.europeana.cloud.service.mcs.persistent.cassandra.CassandraRecordDAO;
 import eu.europeana.cloud.service.mcs.persistent.cassandra.PersistenceUtils;
@@ -284,25 +281,34 @@ public class CassandraDataSetService implements DataSetService {
     }
 
     @Override
+    public List<CompoundDataSetId> getAllDatasetsForRepresentationVersion(Representation representation) throws RepresentationNotExistsException {
+        return new ArrayList<>(
+                dataSetDAO.getDataSetAssignmentsByRepresentationVersion(
+                        representation.getCloudId(),
+                        representation.getRepresentationName(),
+                        representation.getVersion())
+        );
+    }
+
+    @Override
+    public Optional<CompoundDataSetId> getOneDatasetFor(String cloudId, String representationName) throws RepresentationNotExistsException {
+        return dataSetDAO.getOneDataSetFor(cloudId, representationName);
+    }
+
+    @Override
     public void deleteDataSet(String providerId, String dataSetId)
-            throws DataSetNotExistsException {
-        checkIfDatasetExists(dataSetId, providerId);
-        String nextToken = null;
-        int maxSize = 10000;
+            throws DataSetDeletionException, DataSetNotExistsException {
 
-        List<Properties> representations = dataSetDAO.listDataSet(providerId, dataSetId, null, maxSize);
-        for (Properties representation : representations) {
-            removeAssignment(providerId, dataSetId, representation.getProperty("cloudId"), representation.getProperty("schema"), representation.getProperty("versionId"));
+        checkIfDatasetExists(dataSetId,providerId);
+        if (datasetIsEmpty(providerId, dataSetId)) {
+            dataSetDAO.deleteDataSet(providerId, dataSetId);
+        } else {
+            throw new DataSetDeletionException("Can't do it. Dataset is not empty");
         }
+    }
 
-        while (representations.size() == maxSize + 1) {
-            nextToken = representations.get(maxSize).getProperty("nextSlice");
-            representations = dataSetDAO.listDataSet(providerId, dataSetId, nextToken, maxSize);
-            for (Properties representation : representations) {
-                removeAssignment(providerId, dataSetId, representation.getProperty("cloudId"), representation.getProperty("schema"), representation.getProperty("versionId"));
-            }
-        }
-        dataSetDAO.deleteDataSet(providerId, dataSetId);
+    private boolean datasetIsEmpty(String providerId, String dataSetId) {
+        return dataSetDAO.listDataSet(providerId, dataSetId, null, 1).isEmpty();
     }
 
     @Override
@@ -357,7 +363,7 @@ public class CassandraDataSetService implements DataSetService {
         }
     }
 
-    private void checkIfDatasetExists(String dataSetId, String providerId) throws DataSetNotExistsException {
+    public void checkIfDatasetExists(String dataSetId, String providerId) throws DataSetNotExistsException {
         DataSet ds = dataSetDAO.getDataSet(providerId, dataSetId);
         if (ds == null) {
             throw new DataSetNotExistsException();
