+++ service/mcs/rest/src/main/java/eu/europeana/cloud/service/mcs/rest/RepresentationVersionResource.java
@@ -1,27 +1,22 @@
 package eu.europeana.cloud.service.mcs.rest;
 
 import eu.europeana.cloud.common.model.Representation;
-import eu.europeana.cloud.service.aas.authentication.SpringUserUtils;
 import eu.europeana.cloud.service.mcs.RecordService;
-import eu.europeana.cloud.service.mcs.exception.CannotModifyPersistentRepresentationException;
-import eu.europeana.cloud.service.mcs.exception.CannotPersistEmptyRepresentationException;
-import eu.europeana.cloud.service.mcs.exception.RepresentationNotExistsException;
+import eu.europeana.cloud.service.mcs.exception.*;
+import eu.europeana.cloud.service.mcs.utils.DataSetPermissionsVerifier;
 import eu.europeana.cloud.service.mcs.utils.EnrichUriUtil;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.acls.domain.BasePermission;
-import org.springframework.security.acls.domain.ObjectIdentityImpl;
-import org.springframework.security.acls.domain.PrincipalSid;
-import org.springframework.security.acls.model.MutableAcl;
-import org.springframework.security.acls.model.MutableAclService;
-import org.springframework.security.acls.model.ObjectIdentity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 
-import static eu.europeana.cloud.service.mcs.RestInterfaceConstants.*;
+import static eu.europeana.cloud.service.mcs.RestInterfaceConstants.REPRESENTATION_VERSION;
+import static eu.europeana.cloud.service.mcs.RestInterfaceConstants.REPRESENTATION_VERSION_PERSIST;
 
 /**
  * Resource to manage representation versions.
@@ -29,13 +24,16 @@ import static eu.europeana.cloud.service.mcs.RestInterfaceConstants.*;
 @RestController
 public class RepresentationVersionResource {
 
-    private static final String REPRESENTATION_CLASS_NAME = Representation.class.getName();
-    private final RecordService recordService;
-    private final MutableAclService mutableAclService;
+    private static final Logger LOGGER = LoggerFactory.getLogger(RepresentationVersionResource.class.getName());
 
-    public RepresentationVersionResource(RecordService recordService, MutableAclService mutableAclService) {
+    private final RecordService recordService;
+    private final DataSetPermissionsVerifier dataSetPermissionsVerifier;
+
+    public RepresentationVersionResource(
+            RecordService recordService,
+            DataSetPermissionsVerifier dataSetPermissionsVerifier) {
         this.recordService = recordService;
-        this.mutableAclService = mutableAclService;
+        this.dataSetPermissionsVerifier = dataSetPermissionsVerifier;
     }
 
     /**
@@ -51,8 +49,7 @@ public class RepresentationVersionResource {
      * @summary get representation by version
      */
     @GetMapping(value = REPRESENTATION_VERSION, produces = {MediaType.APPLICATION_XML_VALUE, MediaType.APPLICATION_JSON_VALUE})
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version),"
-            + " 'eu.europeana.cloud.common.model.Representation', read)")
+    @PreAuthorize("isAuthenticated()")
     public @ResponseBody Representation getRepresentationVersion(
             HttpServletRequest httpServletRequest,
             @PathVariable String cloudId,
@@ -77,20 +74,19 @@ public class RepresentationVersionResource {
      *                                                       specified version is persistent and as such cannot be removed.
      */
     @DeleteMapping(value = REPRESENTATION_VERSION)
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version)," +
-            " 'eu.europeana.cloud.common.model.Representation', delete)")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteRepresentation(
             @PathVariable String cloudId,
             @PathVariable String representationName,
-            @PathVariable String version) throws RepresentationNotExistsException, CannotModifyPersistentRepresentationException {
+            @PathVariable String version) throws RepresentationNotExistsException, CannotModifyPersistentRepresentationException, AccessDeniedOrObjectDoesNotExistException, DataSetAssignmentException {
 
-        recordService.deleteRepresentation(cloudId, representationName, version);
+        Representation representation = Representation.fromFields(cloudId, representationName, version);
 
-        // let's delete the permissions as well
-        ObjectIdentity dataSetIdentity = new ObjectIdentityImpl(REPRESENTATION_CLASS_NAME,
-                cloudId + "/" + representationName + "/" + version);
-        mutableAclService.deleteAcl(dataSetIdentity, false);
+        if (dataSetPermissionsVerifier.isUserAllowedToDelete(representation)) {
+            recordService.deleteRepresentation(cloudId, representationName, version);
+        }else{
+            throw new AccessDeniedOrObjectDoesNotExistException();
+        }
     }
 
     /**
@@ -111,61 +107,22 @@ public class RepresentationVersionResource {
      * @statuscode 201 representation is made persistent.
      */
     @PostMapping(value = REPRESENTATION_VERSION_PERSIST)
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version),"
-            + " 'eu.europeana.cloud.common.model.Representation', write)")
     public ResponseEntity<Void> persistRepresentation(
             HttpServletRequest httpServletRequest,
             @PathVariable String cloudId,
             @PathVariable String representationName,
             @PathVariable String version) throws RepresentationNotExistsException,
-                        CannotModifyPersistentRepresentationException, CannotPersistEmptyRepresentationException {
+            CannotModifyPersistentRepresentationException, CannotPersistEmptyRepresentationException, AccessDeniedOrObjectDoesNotExistException, DataSetAssignmentException {
 
-        Representation persistentRepresentation = recordService.persistRepresentation(cloudId, representationName, version);
-        EnrichUriUtil.enrich(httpServletRequest, persistentRepresentation);
-        return ResponseEntity.created(persistentRepresentation.getUri()).build();
-    }
+        Representation representation = Representation.fromFields(cloudId, representationName, version);
 
-    /**
-     * Copies all information with all files and their contents from one
-     * representation version to a new temporary one.
-     * <strong>Read permissions required.</strong>
-     *
-     * @param cloudId cloud id of the record which contains the representation version
-     * @param representationName   name of the representation
-     * @param version  a specific version of the representation
-     * @return URI to the created representation in content-location.
-     * @throws RepresentationNotExistsException representation does not exist in
-     *                                          specified version.
-     * @summary copy information including file contents from one representation version to another
-     * @statuscode 201 representation has been copied to a new one.
-     */
-    @PostMapping(value = REPRESENTATION_VERSION_COPY)
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version),"
-            + " 'eu.europeana.cloud.common.model.Representation', read)")
-    public ResponseEntity<Void> copyRepresentation(
-            HttpServletRequest httpServletRequest,
-            @PathVariable String cloudId,
-            @PathVariable String representationName,
-            @PathVariable String version) throws RepresentationNotExistsException {
-
-        Representation representationCopy = recordService.copyRepresentation(cloudId, representationName, version);
-        EnrichUriUtil.enrich(httpServletRequest, representationCopy);
-
-        String copiedReprOwner = SpringUserUtils.getUsername();
-        if (copiedReprOwner != null) {
-            ObjectIdentity versionIdentity = new ObjectIdentityImpl(REPRESENTATION_CLASS_NAME,
-                    cloudId + "/" + representationName + "/" + representationCopy.getVersion());
-
-            MutableAcl versionAcl = mutableAclService.createAcl(versionIdentity);
-
-            versionAcl.insertAce(0, BasePermission.READ, new PrincipalSid(copiedReprOwner), true);
-            versionAcl.insertAce(1, BasePermission.WRITE, new PrincipalSid(copiedReprOwner), true);
-            versionAcl.insertAce(2, BasePermission.DELETE, new PrincipalSid(copiedReprOwner), true);
-            versionAcl.insertAce(3, BasePermission.ADMINISTRATION, new PrincipalSid(copiedReprOwner),true);
-
-            mutableAclService.updateAcl(versionAcl);
+        if (dataSetPermissionsVerifier.isUserAllowedToPersistRepresentation(representation)) {
+            Representation persistentRepresentation = recordService.persistRepresentation(cloudId, representationName, version);
+            EnrichUriUtil.enrich(httpServletRequest, persistentRepresentation);
+            return ResponseEntity.created(persistentRepresentation.getUri()).build();
+        }else{
+            throw new AccessDeniedOrObjectDoesNotExistException();
         }
-
-        return ResponseEntity.created(representationCopy.getUri()).build();
     }
+
 }
